+++ Model/Sync/Td/Delete.php
@@ -14,14 +14,16 @@ class Delete extends \Dotdigitalgroup\Email\Model\Sync\Contact\Delete
         foreach ($collection as $item) {
             $result = true;
             $websiteId = $item->getWebsiteId();
-            $this->_client = $this->_helper->getWebsiteApiClient($websiteId);
-            $importData = unserialize($item->getImportData());
+            if ($this->_helper->isEnabled($websiteId)) {
+                $this->_client = $this->_helper->getWebsiteApiClient($websiteId);
+                $importData = unserialize($item->getImportData());
 
-            if ($this->_client) {
-                $key = $importData[0];
-                $collectionName = $item->getImportType();
-                $this->_client->deleteContactsTransactionalData($key, $collectionName);
-                $this->_handleSingleItemAfterSync($item, $result);
+                if ($this->_client) {
+                    $key = $importData[0];
+                    $collectionName = $item->getImportType();
+                    $this->_client->deleteContactsTransactionalData($key, $collectionName);
+                    $this->_handleSingleItemAfterSync($item, $result);
+                }
             }
         }
     }
